Smart-Ingredient-Analyzer/backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.109.0
uvicorn[standard]==0.27.0

# HTTP client
httpx==0.26.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.0

# CORS middleware (included in FastAPI)
# python-multipart for file uploads (if needed later)
python-multipart==0.0.6

# OCR
pytesseract==0.3.10
Pillow==10.1.0
